Refactor into router, directives and vdom files
